firwmare/canbus: move canmsg defs to another file
